ref(explore): Move updateNullableLocation to utils/url/
