+++ static/app/utils/url/updateNullableLocation.ts
