+++ static/app/views/explore/replays/list/replayQueryParamsProvider.tsx
@@ -2,11 +2,11 @@ import type {ReactNode} from 'react';
 import {useCallback, useMemo} from 'react';
 import type {Location} from 'history';
 
+import {updateNullableLocation} from 'sentry/utils/url/updateNullableLocation';
 import {useLocation} from 'sentry/utils/useLocation';
 import {useNavigate} from 'sentry/utils/useNavigate';
 import {Mode} from 'sentry/views/explore/contexts/pageParamsContext/mode';
 import {QueryParamsContextProvider} from 'sentry/views/explore/queryParams/context';
-import {updateNullableLocation} from 'sentry/views/explore/queryParams/location';
 import {getQueryFromLocation} from 'sentry/views/explore/queryParams/query';
 import {ReadableQueryParams} from 'sentry/views/explore/queryParams/readableQueryParams';
 import {
